+++ docs/source/history.rst
@@ -10,8 +10,8 @@
    unmaintained
 
 
-Unreleased
-==========
+7.6.0
+=====
 
 Features
 --------